new Entity for db: User & Event

+++ src/main/java/com/mentorme/mentor/dto/EventDto.java
@@ -8,8 +8,8 @@ import lombok.Setter;
 public class EventDto {
    
     private Long id;
-    private Integer categoryId;
-    private Integer locationId;
+    private Long categoryId;
+    private Long locationId;
     private String name;
     private String description;
